requirements.txt
async-timeout==4.0.2
blinker==1.6.2
cachelib==0.10.2
click==8.1.3
dnspython==2.3.0
Flask==2.3.2
Flask-Session==0.5.0
Flask-SSE==1.0.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
passlib==1.7.4
pymongo==4.3.3
python-dotenv==1.0.0
redis==4.5.5
six==1.16.0
Werkzeug==2.3.3
zope.event==4.6
zope.interface==6.0
